alternative logic for moving with all arrows
