+++ lua/term-edit/insert.lua
@@ -4,7 +4,7 @@ local async = require 'term-edit.async'
 local config = require 'term-edit.config'
 local M = {}
 
-local function move_keys(len, _)
+local function move_keys(len)
   utils.debug_print('move_by: ', len)
   if len > 0 then
     return string.rep('<Right>', len)
@@ -15,28 +15,17 @@ local function move_keys(len, _)
   end
 end
 
-local function move_keys_up_down(len, lines)
-  utils.debug_print('using up down arrow to move line: ', lines)
-  if lines > 0 then
-    return string.rep('<Down>', lines)
-  elseif lines < 0 then
-    return string.rep('<Up>', -lines)
-  else
-    return move_keys(len, 0)
-  end
-end
-
 ---Enter insert mode and place cursor at target
 ---@param callback? function
 function M.insert_at(target, callback)
   utils.debug_print('enter_insert: target: ', utils.inspect(target))
   local use_up_down_arrows = config.opts.use_up_down_arrows()
   async.vim_cmd('startinsert', function()
-    navigate.navigate_with(
-      target,
-      use_up_down_arrows and move_keys_up_down or move_keys,
-      callback
-    )
+    if use_up_down_arrows then
+      navigate.navigate_all_arrows(target, callback)
+    else
+      navigate.navigate_with(target, move_keys, callback)
+    end
   end)
 end
